Renamed helper file to types in preparation for utils addition.
