+++ src/patient-registration/input/dummy-data/dummy-data-input.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, fireEvent, wait, screen } from '@testing-library/react';
 import { DummyDataInput, dummyFormValues } from './dummy-data-input.component';
 import { initialFormValues } from '../../patient-registration.component';
-import { FormValues } from '../../patient-registration-helper';
+import { FormValues } from '../../patient-registration-types';
 
 describe('dummy data input', () => {
   let formValues: FormValues = initialFormValues;
