+++ src/patient-registration/section/death-info/death-info-section.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { Formik, Form } from 'formik';
 import { initialFormValues } from '../../patient-registration.component';
 import { DeathInfoSection } from './death-info-section.component';
-import { FormValues } from '../../patient-registration-helper';
+import { FormValues } from '../../patient-registration-types';
 
 // TODO: Implement feature and get tests to pass
 describe.skip('death info section', () => {
