+++ src/patient-registration/validation/patient-registration-validation.test.tsx
@@ -6,7 +6,7 @@ import { validationSchema } from './patient-registration-validation';
 import { NameField } from '../field/name/name-field.component';
 import { PatientRegistrationContext } from '../patient-registration-context';
 import { initialFormValues } from '../patient-registration.component';
-import { FormValues } from '../patient-registration-helper';
+import { FormValues } from '../patient-registration-types';
 
 const mockFieldConfigs = {
   name: {
